# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Address:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      postal_code:
        type: string
      province:
        type: string
      type:
        description: home, work, other
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Cart:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: integer
    type: object
  models.CartItem:
    properties:
      cart:
        allOf:
        - $ref: '#/definitions/models.Cart'
        description: Relationships
      cart_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      products:
        description: Relationships
        items:
          $ref: '#/definitions/models.Product'
        type: array
      updated_at:
        type: string
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      order_number:
        type: string
      payment_method:
        type: string
      payment_status:
        description: pending, paid, failed, refunded
        type: string
      shipping_address:
        type: string
      shipping_cost:
        type: integer
      status:
        description: pending, processing, shipped, delivered, cancelled
        type: string
      subtotal:
        type: integer
      tax:
        type: integer
      total_amount:
        type: integer
      tracking_number:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: integer
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order:
        allOf:
        - $ref: '#/definitions/models.Order'
        description: Relationships
      order_id:
        type: integer
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: integer
      unit_price:
        type: integer
      updated_at:
        type: string
    type: object
  models.Product:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      category:
        allOf:
        - $ref: '#/definitions/models.Category'
        description: Relationships
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      dimensions:
        type: string
      id:
        type: integer
      image:
        type: string
      images:
        items:
          type: string
        type: array
      is_active:
        type: boolean
      name:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      price:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
      weight:
        type: number
    type: object
  models.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      addresses:
        description: Relationships
        items:
          $ref: '#/definitions/models.Address'
        type: array
      carts:
        items:
          $ref: '#/definitions/models.Cart'
        type: array
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      phone:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample e-commerce API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1.0"
paths:
  /admin/products:
    post:
      consumes:
      - application/json
      description: Create a new product (admin only)
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new product
      tags:
      - products
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
...
  /products:
    get:
      consumes:
      - application/json
      description: Get a paginated list of products with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search products
        in: query
        name: search
        type: string
      - description: Minimum price filter
        in: query
        name: min_price
        type: integer
      - description: Maximum price filter
        in: query
        name: max_price
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Get a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - products
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"